fix: significant refactor of scratch-to-patch conversion code
