+++ src/conversion/scratch-conversion-data.mjs
@@ -6,18 +6,12 @@ import { indentLines, processInputs } from "./scratch-conversion-helper.mjs";
 export default class ScratchConversionData {
    /**
     * 
-    * @param {Object.<string, ScratchBlock>} blocks 
+    * @param {*} target 
     * @param {string} blockId 
-    * @param {Object.<string, {opcode: string, parameters: string[], exampleParameters: {}}} patchApi 
-    * @param {string[]} patchApiKeys 
-    * @param {*} partialConverter 
-    * @param {*} partialConverterThis 
     * @returns {string}
     */
-   convertDataBlock(blocks, currentBlockId, patchApi, patchApiKeys, partialConverter, partialConverterThis) {
-      const convertBlocksPart = partialConverter.bind(partialConverterThis);
-
-      const currentBlock = blocks[currentBlockId];
+   convertDataBlock(target, currentBlockId) {
+      const currentBlock = target.blocks[currentBlockId];
       const { opcode } = currentBlock;
 
       let script = "";
@@ -25,16 +19,16 @@ export default class ScratchConversionData {
       switch (opcode) {
          case "data_setvariableto": {
             // Set variable
-            const { VARIABLE } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { VALUE } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { VARIABLE } = processInputs(target, currentBlockId, false);
+            const { VALUE } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             script += `${VARIABLE.substring(1, VARIABLE.length - 1)} = ${VALUE}`;
             break;
          }
          case "data_changevariableby": {
             // Change variable by
-            const { VARIABLE } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { VALUE } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { VARIABLE } = processInputs(target, currentBlockId, false);
+            const { VALUE } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             script += `${VARIABLE.substring(1, VARIABLE.length - 1)} += ${VALUE}`;
             break;
@@ -51,8 +45,8 @@ export default class ScratchConversionData {
          }
          case "data_addtolist": {
             // Append to list
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { ITEM } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { LIST } = processInputs(target, currentBlockId, false);
+            const { ITEM } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `${LIST.substring(1, LIST.length - 1)}.append(${ITEM})`;
@@ -60,8 +54,8 @@ export default class ScratchConversionData {
          }
          case "data_deleteoflist": {
             // Delete item at index from list
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { INDEX } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { LIST } = processInputs(target, currentBlockId, false);
+            const { INDEX } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `${LIST.substring(1, LIST.length - 1)}.pop(${INDEX})`;
@@ -69,7 +63,7 @@ export default class ScratchConversionData {
          }
          case "data_deletealloflist": {
             // Clear a list
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
+            const { LIST } = processInputs(target, currentBlockId, false);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `${LIST.substring(1, LIST.length - 1)}.clear()`;
@@ -77,8 +71,8 @@ export default class ScratchConversionData {
          }
          case "data_insertatlist": {
             // Insert an item into the list
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { ITEM, INDEX } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { LIST } = processInputs(target, currentBlockId, false);
+            const { ITEM, INDEX } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `${LIST.substring(1, LIST.length - 1)}.insert(${INDEX}, ${ITEM})`;
@@ -86,8 +80,8 @@ export default class ScratchConversionData {
          }
          case "data_replaceitemoflist": {
             // Replace a list item
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { ITEM, INDEX } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { LIST } = processInputs(target, currentBlockId, false);
+            const { ITEM, INDEX } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `${LIST.substring(1, LIST.length - 1)}[${INDEX}] = ${ITEM}`;
@@ -95,8 +89,8 @@ export default class ScratchConversionData {
          }
          case "data_itemoflist": {
             // Get a list item
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { INDEX } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { LIST } = processInputs(target, currentBlockId, false);
+            const { INDEX } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `${LIST.substring(1, LIST.length - 1)}[${INDEX}]`;
@@ -104,8 +98,8 @@ export default class ScratchConversionData {
          }
          case "data_itemnumoflist": {
             // Get the index of an item in the list
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { ITEM } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { LIST } = processInputs(target, currentBlockId, false);
+            const { ITEM } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `${LIST.substring(1, LIST.length - 1)}.index(${ITEM})`;
@@ -113,7 +107,7 @@ export default class ScratchConversionData {
          }
          case "data_lengthoflist": {
             // Get the length of the list
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
+            const { LIST } = processInputs(target, currentBlockId, false);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `len(${LIST.substring(1, LIST.length - 1)})`;
@@ -121,8 +115,8 @@ export default class ScratchConversionData {
          }
          case "data_listcontainsitem": {
             // Check if the list contains a certain item
-            const { LIST } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false);
-            const { ITEM } = processInputs(blocks, currentBlockId, currentBlock, patchApi, patchApiKeys, convertBlocksPart, false, true);
+            const { LIST } = processInputs(target, currentBlockId, false);
+            const { ITEM } = processInputs(target, currentBlockId, false, true);
             // Add options to change this based on language later.
             console.warn("WARN: using lists as variables isn't currently supported in Patch. Code will be generated but it may or may not function.");
             script += `${ITEM} in ${LIST.substring(1, LIST.length - 1)}`;
